Moved descriptor to the top level
