+++ ledger/ledger-api-bench-tool/src/main/scala/com/daml/ledger/api/benchtool/WorkflowParser.scala
@@ -1,9 +1,9 @@
 // Copyright (c) 2021 Digital Asset (Switzerland) GmbH and/or its affiliates. All rights reserved.
 // SPDX-License-Identifier: Apache-2.0
 
-package com.daml.ledger.api.benchtool.submission
+package com.daml.ledger.api.benchtool
 
-import io.circe._
+import io.circe.Decoder
 import io.circe.yaml.parser
 
 import java.io.Reader
