+++ ledger/ledger-api-bench-tool/src/main/scala/com/daml/ledger/api/benchtool/submission/CommandGenerator.scala
@@ -3,6 +3,7 @@
 
 package com.daml.ledger.api.benchtool.submission
 
+import com.daml.ledger.api.benchtool.SubmissionDescriptor
 import com.daml.ledger.api.v1.commands.Command
 import com.daml.ledger.client.binding.Primitive
 import com.daml.ledger.test.model.Foo._
